added new component to disable local network in settings
