+++ frontend/src/components/SettingsDisableNetworkItem/index.ts
@@ -1,0 +1,1 @@
+export { SettingsDisableNetworkItem } from './SettingsDisableNetworkItem'
